move files out of /wpapi and into root of /lib
