+++ lib/algolia/connector.js
