+++ lib/apolloConfig.js
