+++ lib/frontend/connector.js
@@ -1,6 +1,6 @@
 import {ApolloClient, InMemoryCache} from '@apollo/client'
-import {initializeApollo} from '../apolloConfig'
 import {RestLink} from 'apollo-link-rest'
+import {initializeApollo} from '../apolloConfig'
 
 let feApolloClient
