+++ lib/frontend/wp/archive/queryArchivePosts.js
@@ -1,5 +1,5 @@
+import {wpDataEndpoints} from '@/lib/wordpress/connector'
 import {gql} from '@apollo/client'
-import {wpDataEndpoints} from '@/wpapi/wordpress/connector'
 
 // Query: retrieve archive posts by post type.
 const queryArchivePosts = gql`
