+++ lib/frontend/wp/comments/mutationAddComment.js
@@ -1,6 +1,6 @@
+import {wpDataEndpoints} from '@/lib/wordpress/connector'
+import commentsFields from '@/lib/wordpress/_partials/commentsFields'
 import {gql} from '@apollo/client'
-import {wpDataEndpoints} from '@/wpapi/wordpress/connector'
-import commentsFields from '@/wpapi/wordpress/_partials/commentsFields'
 
 // Mutation: Add a comment to the given post
 const mutationAddComment = gql`
