+++ lib/frontend/wp/comments/postComment.js
