+++ lib/wordpress/_global/getFrontendPage.js
