+++ lib/wordpress/_global/getPostTypeArchive.js
