+++ lib/wordpress/_global/getPostTypeById.js
