+++ lib/wordpress/_global/getPostTypeStaticPaths.js
