+++ lib/wordpress/_global/getPostTypeStaticProps.js
@@ -1,12 +1,12 @@
-import {algoliaIndexName} from '@/wpapi/algolia/connector'
-import getPostTypeById from './getPostTypeById'
-import getPostTypeArchive, {archiveQuerySeo} from './getPostTypeArchive'
-import {addApolloState} from '@/wpapi/apolloConfig'
+import {algoliaIndexName} from '@/lib/algolia/connector'
+import {addApolloState} from '@/lib/apolloConfig'
 import getFrontendPage, {frontendPageSeo} from './getFrontendPage'
+import getPostTypeArchive, {archiveQuerySeo} from './getPostTypeArchive'
+import getPostTypeById from './getPostTypeById'
+import getPostTypeTaxonomyArchive from './getPostTypeTaxonomyArchive'
 import getSettingsCustomPage, {
   customPageQuerySeo
 } from './getSettingsCustomPage'
-import getPostTypeTaxonomyArchive from './getPostTypeTaxonomyArchive'
 
 /**
  * Retrieve static props by post type.
