+++ lib/wordpress/_global/getPostTypeTaxonomyArchive.js
