+++ lib/wordpress/_global/getSettingsCustomPage.js
