+++ lib/wordpress/_global/postCommentToPost.js
