+++ lib/wordpress/_global/postTypes.js
