+++ lib/wordpress/_global/processPostTypeQuery.js
