+++ lib/wordpress/_global/queryDefaultPageData.js
