+++ lib/wordpress/_partials/archivePageInfo.js
