+++ lib/wordpress/_partials/categoriesPostFields.js
