+++ lib/wordpress/_partials/featuredImagePostFields.js
