+++ lib/wordpress/_partials/globalPostFields.js
