+++ lib/wordpress/connector.js
