+++ lib/wordpress/gravityForms/queryFormById.js
