+++ lib/wordpress/media/queryMediaAttributes.js
